+++ src/monitor.h
@@ -122,7 +122,7 @@ class Monitor {
   // sampling is enabled. locking_method_ may be null if the lock is currently
   // unlocked, or if the lock is acquired by the system when the stack is empty.
   const Method* locking_method_;
-  uint32_t locking_pc_;
+  uintptr_t locking_pc_;
 
   friend class MonitorList;
   friend class Object;
